Python

+++ sdk/python/pulumi_azure_nextgen/avs/v20200717preview/workload_network_dhcp.py
@@ -21,7 +21,7 @@ class WorkloadNetworkDhcp(pulumi.CustomResource):
                  display_name: Optional[pulumi.Input[str]] = None,
                  private_cloud_name: Optional[pulumi.Input[str]] = None,
                  resource_group_name: Optional[pulumi.Input[str]] = None,
-                 revision: Optional[pulumi.Input[int]] = None,
+                 revision: Optional[pulumi.Input[float]] = None,
                  __props__=None,
                  __name__=None,
                  __opts__=None):
@@ -35,7 +35,7 @@ class WorkloadNetworkDhcp(pulumi.CustomResource):
         :param pulumi.Input[str] display_name: Display name of the DHCP entity.
         :param pulumi.Input[str] private_cloud_name: Name of the private cloud
         :param pulumi.Input[str] resource_group_name: The name of the resource group. The name is case insensitive.
-        :param pulumi.Input[int] revision: NSX revision number.
+        :param pulumi.Input[float] revision: NSX revision number.
         """
         if __name__ is not None:
             warnings.warn("explicit use of __name__ is deprecated", DeprecationWarning)
@@ -130,7 +130,7 @@ class WorkloadNetworkDhcp(pulumi.CustomResource):
 
     @property
     @pulumi.getter
-    def revision(self) -> pulumi.Output[Optional[int]]:
+    def revision(self) -> pulumi.Output[Optional[float]]:
         """
         NSX revision number.
         """
