+++ sdk/python/pulumi_azure_nextgen/avs/v20200717preview/workload_network_dns_service.py
@@ -24,7 +24,7 @@ class WorkloadNetworkDnsService(pulumi.CustomResource):
                  log_level: Optional[pulumi.Input[Union[str, 'DnsServiceLogLevelEnum']]] = None,
                  private_cloud_name: Optional[pulumi.Input[str]] = None,
                  resource_group_name: Optional[pulumi.Input[str]] = None,
-                 revision: Optional[pulumi.Input[int]] = None,
+                 revision: Optional[pulumi.Input[float]] = None,
                  __props__=None,
                  __name__=None,
                  __opts__=None):
@@ -41,7 +41,7 @@ class WorkloadNetworkDnsService(pulumi.CustomResource):
         :param pulumi.Input[Union[str, 'DnsServiceLogLevelEnum']] log_level: DNS Service log level.
         :param pulumi.Input[str] private_cloud_name: Name of the private cloud
         :param pulumi.Input[str] resource_group_name: The name of the resource group. The name is case insensitive.
-        :param pulumi.Input[int] revision: NSX revision number.
+        :param pulumi.Input[float] revision: NSX revision number.
         """
         if __name__ is not None:
             warnings.warn("explicit use of __name__ is deprecated", DeprecationWarning)
@@ -161,7 +161,7 @@ class WorkloadNetworkDnsService(pulumi.CustomResource):
 
     @property
     @pulumi.getter
-    def revision(self) -> pulumi.Output[Optional[int]]:
+    def revision(self) -> pulumi.Output[Optional[float]]:
         """
         NSX revision number.
         """
