+++ sdk/python/pulumi_azure_nextgen/avs/v20200717preview/workload_network_dns_zone.py
@@ -17,12 +17,12 @@ class WorkloadNetworkDnsZone(pulumi.CustomResource):
                  opts: Optional[pulumi.ResourceOptions] = None,
                  display_name: Optional[pulumi.Input[str]] = None,
                  dns_server_ips: Optional[pulumi.Input[Sequence[pulumi.Input[str]]]] = None,
-                 dns_services: Optional[pulumi.Input[int]] = None,
+                 dns_services: Optional[pulumi.Input[float]] = None,
                  dns_zone_id: Optional[pulumi.Input[str]] = None,
                  domain: Optional[pulumi.Input[Sequence[pulumi.Input[str]]]] = None,
                  private_cloud_name: Optional[pulumi.Input[str]] = None,
                  resource_group_name: Optional[pulumi.Input[str]] = None,
-                 revision: Optional[pulumi.Input[int]] = None,
+                 revision: Optional[pulumi.Input[float]] = None,
                  source_ip: Optional[pulumi.Input[str]] = None,
                  __props__=None,
                  __name__=None,
@@ -34,12 +34,12 @@ class WorkloadNetworkDnsZone(pulumi.CustomResource):
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[str] display_name: Display name of the DNS Zone.
         :param pulumi.Input[Sequence[pulumi.Input[str]]] dns_server_ips: DNS Server IP array of the DNS Zone.
-        :param pulumi.Input[int] dns_services: Number of DNS Services using the DNS zone.
+        :param pulumi.Input[float] dns_services: Number of DNS Services using the DNS zone.
         :param pulumi.Input[str] dns_zone_id: NSX DNS Zone identifier. Generally the same as the DNS Zone's display name
         :param pulumi.Input[Sequence[pulumi.Input[str]]] domain: Domain names of the DNS Zone.
         :param pulumi.Input[str] private_cloud_name: Name of the private cloud
         :param pulumi.Input[str] resource_group_name: The name of the resource group. The name is case insensitive.
-        :param pulumi.Input[int] revision: NSX revision number.
+        :param pulumi.Input[float] revision: NSX revision number.
         :param pulumi.Input[str] source_ip: Source IP of the DNS Zone.
         """
         if __name__ is not None:
@@ -119,7 +119,7 @@ class WorkloadNetworkDnsZone(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="dnsServices")
-    def dns_services(self) -> pulumi.Output[Optional[int]]:
+    def dns_services(self) -> pulumi.Output[Optional[float]]:
         """
         Number of DNS Services using the DNS zone.
         """
@@ -151,7 +151,7 @@ class WorkloadNetworkDnsZone(pulumi.CustomResource):
 
     @property
     @pulumi.getter
-    def revision(self) -> pulumi.Output[Optional[int]]:
+    def revision(self) -> pulumi.Output[Optional[float]]:
         """
         NSX revision number.
         """
